comprobación de si el articulo está o no en db

+++ app/src/main/java/com/example/wikimvvm/viewmodel/ArticleViewModel.kt
@@ -1,6 +1,7 @@
 package com.example.wikimvvm.viewmodel
 
 import android.content.Context
+import android.widget.Toast
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.room.Room
@@ -34,25 +35,35 @@ class ArticleViewModel(context: Context) : ViewModel() {
         }
     }
 
-    fun checkIfArticleInDb(articleResponse: ArticleResponse) {
+    fun addArticleToFavourite(toast: Toast, articleResponse: ArticleResponse){
         CoroutineScope(Dispatchers.IO).launch {
             if (checkIfArticleInFavouritesUseCase.checkIfArticleInFavourite(articleResponse)) {
                 insertFavouriteArticleUseCase.insertFavouriteArticle(articleResponse)
+                toast.show()
+            } else {
+                toast.setText("${articleResponse.title} ya está en favoritos")
+                toast.show()
             }
         }
     }
 
-    fun emptyListOfFavourites() {
+    fun emptyListOfFavourites(toast: Toast) {
         CoroutineScope(Dispatchers.IO).launch {
+            toast.show()
             emptyListOfFavouritesUseCase.emptyListOfFavourites()
         }
     }
 
-    fun deleteArticleFromFavourites(articleToDelete: ArticleResponse) {
+    fun deleteArticleFromFavourites(toast: Toast, articleToDelete: ArticleResponse) {
         CoroutineScope(Dispatchers.IO).launch {
-            deleteArticleFromFavouritesUseCase.deleteArticleFromFavourites(articleToDelete)
+            if (!checkIfArticleInFavouritesUseCase.checkIfArticleInFavourite(articleToDelete)){
+                toast.show()
+                deleteArticleFromFavouritesUseCase.deleteArticleFromFavourites(articleToDelete)
+            } else {
+                toast.setText("${articleToDelete.title} no está en favoritos")
+                toast.show()
+            }
         }
     }
-
 }
